Added solution file and changed the comment style
In particular, added the file rosalindSolutions.py, which is
set up to compute the solutions to some slightly larger data sets
that uses the algorithms in the other files. It is mainly set up
to read from text files in a certain format, parse the data, and then
compute the solution to a given problem

I also debugged and solidified my function to find the shortest
cycle between a given edge of a directed graph. If you consider
the edge a --> b in a graph, then finding the shortest cycle through
a --> b is isomorphic to the problem of performing a shortest path
search from b --> a in the graph sans the edge a --> b.